Moved filters to own Package
